Conflicts:
	heap.h
	solve_nine.c
	test_solve.c

+++ heap.h
@@ -1,16 +1,16 @@
 #ifndef HEAP_LIB
 #define HEAP_LIB
-typedef long (*HEAP_CMP_FUNC)(void **, long, long);
-typedef void (*HEAP_NODE_CALLBACK)(void *, long);
+typedef int (*HEAP_CMP_FUNC)(void **, int, int);
+typedef void (*HEAP_NODE_CALLBACK)(void *, int);
 #endif
 
-void alloc_heap(void ***, long);
-void build_min_heap(void **, long, HEAP_CMP_FUNC, HEAP_NODE_CALLBACK);
+void alloc_heap(void ***, int);
+void build_min_heap(void **, int, HEAP_CMP_FUNC, HEAP_NODE_CALLBACK);
 void *heap_minium(void **);
 void heap_extract_min(void **, HEAP_CMP_FUNC, HEAP_NODE_CALLBACK);
 void min_heap_insert(void ***, void *, HEAP_CMP_FUNC, HEAP_NODE_CALLBACK);
-void min_heap_decrease_key(void **, long, HEAP_CMP_FUNC, HEAP_NODE_CALLBACK);
-long heap_not_empty(void);
+void min_heap_decrease_key(void **, int, HEAP_CMP_FUNC, HEAP_NODE_CALLBACK);
+int heap_not_empty(void);
 
-extern long array_size;
-extern long heap_size;
+extern int array_size; 
+extern int heap_size;
